safari-600.7-branch/LayoutTests/fast/viewport/viewport-33-expected.txt

ALERT: viewport size 320x352 scale 2 with limits [1, 5] and userScalable true
